Add more view to title routing to correspond to current position in terms of current tab and navigation route. Also clear bug with empty friend list.
The empty friend list bug was that if there was dummy data in the state's
friendList, and the get fetch does not return a real friend list, then the
user may be clicking on dummy data, expecting it to be a real friend. This
dummy won't have the right params to query the db, causing an error.