Accept and reject friends updates the pending friend req and friendslist
Note the commented out setTimeout in the requestRow. This was to set a delay in popping the navigation back to the friends tab view, so that
the user could see some sort of "Request Sent" notificiation. This would
require there being some state in the row component -- I'm holding
off on that for now.