Set up client-side location.
The geocoder is in the correct place, and the code preceding it is correctly grabbing data from the device. The geocoder needs Xcode
config in order to work. Dummy data is being used to establish
all client-side logic dealing with the location. The server still needs
too be updated to accept the position in new entries.